ov10635: report all supported frame sizes
Current implementation reports only one frame size 1280x800. This is
the best resolution which is supported by the device. But according
to the datasheet, the device supports the following list of
framesizes:
 - 1280x800 (WXGA);
 - 1280x720 (HD720p);
 -  752x480 (WVGA);
 -  640x480 (VGA);
 -  600x400;
 -  352x288 (CIF);
 -  320x240 (QVGA).

The possible solution is to report all frame sizes from the list with
`V4L2_FRMSIZE_TYPE_DISCRETE` type. But taking into account that
current driver implementation supports cropping it is proposed to
report that `V4L2_FRMSIZE_TYPE_CONTINUOUS` supported with width
320/1280 and height 240/800.

Mantis-Id: 0005883
Change-Id: I796b7d112e8feba13db0caf3dba8f3d54128b662
Signed-off-by: Sergii Piatakov <sergii.piatakov@globallogic.com>