RTC: Add early_param to set initial time (rtc-fake)
Added ability to set the initial time of this fake rtc
by the Kernel bootargs (early_param).
To set the time it's needed to add to bootargs the following:
init_time=N, where N is time in seconds e.g.
init_time=1475750706 sets time to Oct 06, 2016 - 10:45:06
If the early_param isn't set, the default time is
Sep 26, 2016 - 15:23:07 (defined in the driver code).

Change-Id: I2946ebaeb674b3c2fcc6bbef75ddb1247ff8d1c2
Signed-off-by: Oleg Kosheliev <oleg.kosheliev@globallogic.com>
Signed-off-by: Sergii Boryshchenko <sergii.boryshchenko@globallogic.com>