defconfig: pass max partitions into the loop module
The loop module has a `module_param` which specifies the maximum
number of partitions on the loop device. Based on this value the
module decides to apply or not the following flags `LO_FLAGS_PARTSCAN`
and `GENHD_FL_NO_PART_SCAN`. The flags may affect the `BLKRRPART`
ioctl which is used to reread a partition table in the runtime.

This functionality is required by the `sm` Android command when it
tries to set up a virtual disk and calls `sgdisk` utility to sync
disk info.

Test: try on the DUT the following scenario:
      $ sm set-virtual-disk true
      $ sm list-disks
      $ disk:7,0
      $ sm partition disk:7,0 public
      $ sm list-volumes
      $ public:7,1 mounted ABCD-1234
      or run CTS:
      cts-tradefed run commandAndExit cts-dev \
      -m CtsAppSecurityHostTestCases
Jira-Id: ADM-2252
Change-Id: Id628a49e79d44feb4811cedb107111ed3353ae2c
Signed-off-by: Sergii Piatakov <sergii.piatakov@globallogic.com>