xhci: Add XHCI_TRUST_TX_LENGTH quirk
This patch fixes warning "WARN Successful completion on short 
TX for slot 1 ep 2: needs XHCI_TRUST_TX_LENGTH quirk?"

Description of the problem:

The successful completion status tells us the data from the
device was received fine, but then the "untransferred length" field says
that some of the data wasn't received.  On a short transfer from the
device, the host is supposed to pass a short transfer completion code.

With the XHCI_TRUST_TX_LENGTH quirk, we will trust the untransferred
length, not the completion status.

Jira-Id: ADM-3050

Change-Id: Ibd8622646c788956bc351ca854db8d5a4ef09f41
Signed-off-by: Andrii Shevchenko <andrii.k.shevchenko@globallogic.com>