soc/camera: ordering camera ports initialization
The original implementation of the camera host registration function
didn't take into account order in which camera ports were declared in
dts files. That led to the kernel crash at the time of the rcar vin
probing. So we've changed the search function to keep the order
of the ports.

Jira-Id: ADM-3103
Change-Id: I3f89c098dc2842d2cf8399c90a8dd7ddf05ca760
Signed-off-by: Sergii Boryshchenko <sergii.boryshchenko@globallogic.com>