ANDROID: ext4: Handle casefolding with encryption
This adds support for encryption with casefolding.

Since the name on disk is case preserving, and also encrypted, we can no
longer just recompute the hash on the fly. Additionally, to avoid
leaking extra information from the hash of the unencrypted name, we use
siphash via an fscrypt v2 policy.

The hash is stored at the end of the directory entry for all entries
inside of an encrypted and casefolded directory apart from those that
deal with '.' and '..'. This way, the change is backwards compatible
with existing ext4 filesystems.

Signed-off-by: Daniel Rosenberg <drosen@google.com>
Test: Boots, /data/media is case insensitive
Bug: 138322712
Jira-Id: ADM-5213
Test: ensure that logcat doesn't contain:
EXT4-fs (mmcblk0p13): Can't mount with encoding and encryption
Signed-off-by: Sergii Piatakov <sergii.piatakov@globallogic.com>
Change-Id: I07354e3129aa07d309fbe36c002fee1af718f348