rcar-vin: align supported formats
According to the documentation, some channels do not support all output
formats. Namely, channels 2 and 3 do not support output NV12 format.
But the current version of the driver reports the same list of formats
for all channels. Such behavior does not allow to distinguish between
cameras with and without NV12 support in the userspace. So it is
proposed to report that NV12 format is supported only for channels 0
and 1.

Test: one possible test case is: build `v4l-utils` and then check
      supported formats for all cameras on the DUT using the following
      command:
      # v4l2-ctl -d /dev/videoX --list-firmats
Jira-Id: ADM-3282
Change-Id: I4e2c2a9d1e899230fb12674e9c6cbaf52f8a57a5
Signed-off-by: Sergii Piatakov <sergii.piatakov@globallogic.com>