bugfix: f_fs: Fix double free issue
VtsHalUsbGadgetV1_0Host test caused the kernel crash
due to double free when the same work was scheduled
for completion by ffs_epfile_async_io_complete() and
then ffs_aio_cancel() was called right before the WQ
executed. This patch checks if the same work was
already scheduled and cancels it or skips submission
if the first work was already started.

Jira-Id: ADM-4115
Signed-off-by: Vitaliy Vasylskyy <vitaliy.vasylskyy@globallogic.com>
Change-Id: I7ea21852b6f7bdb60e102de14d964ba737d3133d