Android.mk: split DTBs depending on target build platform
Currently, the dtb image for M3 or M3N contains dtb and overlays
for H3 platform, which redundant. They can be split in build time
depending on the target platform. This reduces total dtb and dtbo
images size, time for avb verification in U-Boot and possible issues
due to applying overlays in U-Boot for improper platform.

Jira-Id: ADM-5030
Change-Id: Iefe00eb64c4e30a325843c72448df7830a3ddeac
Signed-off-by: Leonid Komaryanskiy <leonid.komaryanskiy@globallogic.com>